+++ apps/mark-scan/frontend/src/pages/admin_screen.test.tsx
@@ -12,7 +12,7 @@ import { election } from '../../test/helpers/election';
 import { advanceTimers } from '../../test/helpers/timers';
 
 import { AdminScreen, AdminScreenProps } from './admin_screen';
-import { mockMachineConfig } from '../../test/helpers/fake_machine_config';
+import { mockMachineConfig } from '../../test/helpers/mock_machine_config';
 import {
   ApiMock,
   createApiMock,
