+++ apps/mark-scan/frontend/test/helpers/mock_api_client.tsx
@@ -31,7 +31,7 @@ import { err, ok, Result } from '@votingworks/basics';
 import type { BatteryInfo, DiskSpaceSummary } from '@votingworks/backend';
 import { TestErrorBoundary } from '@votingworks/ui';
 import type { UsbDriveStatus } from '@votingworks/usb-drive';
-import { mockMachineConfig } from './fake_machine_config';
+import { mockMachineConfig } from './mock_machine_config';
 import { initialElectionState } from '../../src/app_root';
 import { ApiProvider } from '../../src/api_provider';
