+++ apps/mark-scan/frontend/test/helpers/mock_auth.ts
