+++ apps/mark-scan/frontend/test/helpers/mock_machine_config.ts
