+++ apps/mark-scan/frontend/test/test_utils.tsx
@@ -15,7 +15,7 @@ import { electionGeneralDefinition } from '@votingworks/fixtures';
 import { render as testRender } from './react_testing_library';
 
 import { BallotContext } from '../src/contexts/ballot_context';
-import { mockMachineConfig } from './helpers/fake_machine_config';
+import { mockMachineConfig } from './helpers/mock_machine_config';
 import { ApiMock, createApiMock } from './helpers/mock_api_client';
 import { ApiProvider } from '../src/api_provider';
