+++ apps/mark/frontend/src/pages/admin_screen.test.tsx
@@ -15,7 +15,7 @@ import { election, defaultPrecinctId } from '../../test/helpers/election';
 import { advanceTimers } from '../../test/helpers/timers';
 
 import { AdminScreen, AdminScreenProps } from './admin_screen';
-import { mockMachineConfig } from '../../test/helpers/fake_machine_config';
+import { mockMachineConfig } from '../../test/helpers/mock_machine_config';
 import {
   ApiMock,
   createApiMock,
