+++ apps/mark/frontend/src/pages/poll_worker_screen.test.tsx
@@ -28,8 +28,8 @@ import { render } from '../../test/test_utils';
 import { defaultPrecinctId } from '../../test/helpers/election';
 
 import { PollWorkerScreen, PollworkerScreenProps } from './poll_worker_screen';
-import { mockMachineConfig } from '../../test/helpers/fake_machine_config';
-import { mockDevices } from '../../test/helpers/fake_devices';
+import { mockMachineConfig } from '../../test/helpers/mock_machine_config';
+import { mockDevices } from '../../test/helpers/mock_devices';
 import { ApiMock, createApiMock } from '../../test/helpers/mock_api_client';
 import { ApiProvider } from '../api_provider';
