+++ apps/mark/frontend/test/helpers/mock_machine_config.ts
