+++ libs/test-utils/src/index.ts
@@ -6,10 +6,10 @@ export * from './child_process';
 export * from './compressed_tallies';
 export * from './console';
 export * from './expect_print';
-export * from './fake_file_writer';
-export * from './fake_kiosk';
-export * from './fake_printer';
-export * from './fake_use_audio_controls';
+export * from './mock_file_writer';
+export * from './mock_kiosk';
+export * from './mock_printer';
+export * from './mock_use_audio_controls';
 export * from './has_text_across_elements';
 export * from './mock_function';
 export * from './mock_of';
