+++ libs/test-utils/src/mock_kiosk.ts
